+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
   * Fixed incorrent unix socket path length check (gh-341).
+  * Now net_box_uri can be accepted as table (gh-342).
 
 ## 1.0.0